remove unused declaration

+++ src/runtime/object.cpp
@@ -1578,7 +1578,7 @@ extern "C" LEAN_EXPORT usize lean_usize_big_modn(usize a1, b_lean_obj_arg) {
 // IntX
 
 extern "C" LEAN_EXPORT int8 lean_int8_of_big_int(b_obj_arg a) {
-    return static_cast<int8>(mpz_value(a).smod8());
+    return mpz_value(a).smod8();
 }
 
 // =======================================
